refactoring state.js
