+++ packages/vtv/src/model/util/index.js
@@ -7,3 +7,4 @@ export { default as getState } from './getState'
 export { default as getChildState } from './getChildState'
 export { default as getNestedState } from './getNestedState'
 export { default as getDraftUpdate } from './getDraftUpdate'
+export { default as removeTemporaryState } from './removeTemporaryState'
